requirements.txt
django
django-celery
django-jsonfield
django-auth-ldap  
python-magic
South==0.7.6
simplejson
#pil
pexpect
django-guardian
celery
MySQL-python
python-ldap
